+++ cpp/src/sdk/examples/simple/src/ClientTest.cpp
@@ -4,13 +4,13 @@
  * Proprietary and confidential.
  ******************************************************************************/
 #include "ClientTest.h"
-#include "MegaSearch.h"
+#include "MilvusApi.h"
 
 #include <iostream>
 #include <time.h>
 #include <unistd.h>
 
-using namespace megasearch;
+using namespace ::milvus;
 
 namespace {
     std::string GetTableName();
@@ -20,15 +20,16 @@ namespace {
     static constexpr int64_t TOTAL_ROW_COUNT = 100000;
     static constexpr int64_t TOP_K = 10;
     static constexpr int64_t SEARCH_TARGET = 5000; //change this value, result is different
+    static constexpr int64_t ADD_VECTOR_LOOP = 1;
 
 #define BLOCK_SPLITER std::cout << "===========================================" << std::endl;
 
-    void PrintTableSchema(const megasearch::TableSchema& tb_schema) {
+    void PrintTableSchema(const TableSchema& tb_schema) {
         BLOCK_SPLITER
         std::cout << "Table name: " << tb_schema.table_name << std::endl;
         std::cout << "Table index type: " << (int)tb_schema.index_type << std::endl;
         std::cout << "Table dimension: " << tb_schema.dimension << std::endl;
-        std::cout << "Table store raw data: " << tb_schema.store_raw_vector << std::endl;
+        std::cout << "Table store raw data: " << (tb_schema.store_raw_vector ? "true" : "false") << std::endl;
         BLOCK_SPLITER
     }
 
@@ -71,6 +72,18 @@ namespace {
         std::string str = std::to_string(t->tm_year + 1900) + "_" + std::to_string(t->tm_mon + 1)
                           + "_" + std::to_string(t->tm_mday) + "_" + std::to_string(t->tm_hour)
                           + "_" + std::to_string(t->tm_min) + "_" + std::to_string(t->tm_sec);
+
+        return str;
+    }
+
+    std::string CurrentTmDate() {
+        time_t tt;
+        time( &tt );
+        tt = tt + 8*3600;
+        tm* t= gmtime( &tt );
+
+        std::string str = std::to_string(t->tm_year + 1900) + "-" + std::to_string(t->tm_mon + 1)
+                          + "-" + std::to_string(t->tm_mday);
 
         return str;
     }
@@ -122,7 +135,7 @@ ClientTest::Test(const std::string& address, const std::string& port) {
 
     {//server version
         std::string version = conn->ServerVersion();
-        std::cout << "MegaSearch server version: " << version << std::endl;
+        std::cout << "Server version: " << version << std::endl;
     }
 
     {//sdk version
@@ -136,15 +149,17 @@ ClientTest::Test(const std::string& address, const std::string& port) {
         std::cout << "ShowTables function call status: " << stat.ToString() << std::endl;
         std::cout << "All tables: " << std::endl;
         for(auto& table : tables) {
-            std::cout << "\t" << table << std::endl;
+            int64_t row_count = 0;
+            stat = conn->GetTableRowCount(table, row_count);
+            std::cout << "\t" << table << "(" << row_count << " rows)" << std::endl;
         }
     }
 
     {//create table
         TableSchema tb_schema = BuildTableSchema();
-        PrintTableSchema(tb_schema);
         Status stat = conn->CreateTable(tb_schema);
         std::cout << "CreateTable function call status: " << stat.ToString() << std::endl;
+        PrintTableSchema(tb_schema);
     }
 
     {//describe table
@@ -154,9 +169,9 @@ ClientTest::Test(const std::string& address, const std::string& port) {
         PrintTableSchema(tb_schema);
     }
 
-    {//add vectors
+    for(int i = 0; i < ADD_VECTOR_LOOP; i++){//add vectors
         std::vector<RowRecord> record_array;
-        BuildVectors(0, TOTAL_ROW_COUNT, record_array);
+        BuildVectors(i*TOTAL_ROW_COUNT, (i+1)*TOTAL_ROW_COUNT, record_array);
         std::vector<int64_t> record_ids;
         Status stat = conn->AddVector(TABLE_NAME, record_array, record_ids);
         std::cout << "AddVector function call status: " << stat.ToString() << std::endl;
@@ -170,6 +185,10 @@ ClientTest::Test(const std::string& address, const std::string& port) {
         BuildVectors(SEARCH_TARGET, SEARCH_TARGET + 10, record_array);
 
         std::vector<Range> query_range_array;
+        Range rg;
+        rg.start_value = CurrentTmDate();
+        rg.end_value = CurrentTmDate();
+        query_range_array.emplace_back(rg);
         std::vector<TopKQueryResult> topk_query_result_array;
         Status stat = conn->SearchVector(TABLE_NAME, record_array, query_range_array, TOP_K, topk_query_result_array);
         std::cout << "SearchVector function call status: " << stat.ToString() << std::endl;