+++ cpp/src/sdk/src/interface/ConnectionImpl.cpp
@@ -4,8 +4,9 @@
  * Proprietary and confidential.
  ******************************************************************************/
 #include "ConnectionImpl.h"
+#include "version.h"
 
-namespace megasearch {
+namespace milvus {
 
 std::shared_ptr<Connection>
 Connection::Create() {
@@ -13,7 +14,7 @@ Connection::Create() {
 }
 
 Status
-Connection::Destroy(std::shared_ptr<megasearch::Connection> connection_ptr) {
+Connection::Destroy(std::shared_ptr<milvus::Connection> connection_ptr) {
     if(connection_ptr != nullptr) {
         return connection_ptr->Disconnect();
     }
@@ -47,7 +48,7 @@ ConnectionImpl::Disconnect() {
 
 std::string
 ConnectionImpl::ClientVersion() const {
-    return client_proxy_->ClientVersion();
+    return MILVUS_VERSION;
 }
 
 Status