+++ cpp/src/wrapper/Operand.h
@@ -13,7 +13,7 @@
 
 
 namespace zilliz {
-namespace vecwise {
+namespace milvus {
 namespace engine {
 
 struct Operand {