+++ cpp/start_server.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 
-./cmake_build/src/vecwise_server -c ./conf/server_config.yaml -l ./conf/vecwise_engine_log.conf &
+./cmake_build/src/milvus_server -c ./conf/server_config.yaml -l ./conf/log_config.conf &